Update issue #17 Improved Logal Signal Strength Gradient algorithm by weighting each calculated "arrow" by the signal strength of the center point.
git-svn-id: https://wificompass.googlecode.com/svn/trunk@95 6f3e661d-c991-e48d-a97c-fec7ba324fcc

+++ src/at/fhstp/wificompass/triangulation/AccessPointTriangulation.java
@@ -112,7 +112,7 @@ public abstract class AccessPointTriangulation {
 			}
 			
 			count ++;
-			Logger.d(count / measurementData.size() * 100 + " % (" + count + " out of " + measurementData.size() + ") done.");
+			Logger.d((float)count / measurementData.size() * 100 + " % (" + count + " out of " + measurementData.size() + ") done.");
 		}
 
 		return aps;
